added apartment_sponsorship migration

+++ database/migrations/2023_03_22_161543_create_sponsorships_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('sponsorships', function (Blueprint $table) {
             $table->id();
             $table->string('Titolo');
-            $table->float('Prezzo',3,2);
+            $table->decimal('Prezzo',3,2);
             $table->UnsignedTinyInteger('Ore_valide');
             $table->timestamps();
         });
